Create rehype-twemoji plugin & Restructure files
NOTE: The plugin rehype-twemoji is still unused as the current
implementation uses jsdelivr to deliver SVG images of emoji.
Usage of this plugin may result in violation of GPDR laws
because user's IP address gets leaked to jsdelivr website
without prior knowledge or consent from the user.

Need to find a way for this rehype-twemoji plugin and Astro's Vite
to work together with locally stored SVG files from twemoji project.